add tests create products and findAll

+++ internal/entity/product.go
@@ -18,11 +18,11 @@ var (
 type Product struct {
 	ID        entity.ID `json:"id"`
 	Name      string    `json:"name"`
-	Price     int       `json:"price"`
+	Price     float64   `json:"price"`
 	CreatedAt time.Time `json:"created_at"`
 }
 
-func NewProduct(name string, price int) (*Product, error) {
+func NewProduct(name string, price float64) (*Product, error) {
 	product := &Product{
 		ID:        entity.NewID(),
 		Name:      name,
@@ -46,7 +46,7 @@ func (p *Product) Validate() error {
 	if p.Name == "" {
 		return ErrNameIsRequired
 	}
-	if p.Price == 0 {
+	if p.Price == 0.0 {
 		return ErrPriceIsRequired
 	}
 	if p.Price < 0 {
